refactor: overlappingDays() return type from Double to long

+++ src/Accouting.java
@@ -20,6 +20,6 @@ public class Accouting {
 		Period period = new Period(start, end);
 		Budget budget = budgets.get(0);
 
-		return period.overlappingDays(budget);
+		return (double) period.overlappingDays(budget);
 	}
 }
